misc/HtmlTestSuite/results/test-0823.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
  <meta charset="utf-8" id="expected"/>
  <title>Selectors: attribute selectors with missing right token</title>
  <link href="mailto:ericwilligers@chromium.org" rel="author" title="Eric Willigers"/>
  <link href="https://drafts.csswg.org/selectors-4/#attribute-selectors" rel="help"/>
  <link href="https://drafts.csswg.org/css-syntax/#rule-defs" rel="help"/>
  <meta content="Checks attribute selectors with missing right token are matched." name="assert"/>
  <script src="/resources/testharness.js"/>
  <script src="/resources/testharnessreport.js"/>
</head>
<body>


  <div id="container">
    <span/>
    <span class="cls"/>
  </div>

  <script>
    test(() =&gt; {
      assert_equals(document.querySelector('meta[charset="utf-8"'), expected);
      assert_equals(document.querySelector('meta[charset="utf-8'), expected);

      assert_equals(container.querySelectorAll('span:not([class]').length, 1);
      assert_equals(container.querySelectorAll('span:not([class').length, 1);
    }, "attribute selectors with missing right tokens succeed");
  </script>
</body>
</html>
